feat: minor data refactors
